🔨 refactor: Replace `with-eval-after-load` with `mazd//after` macro
Introduce the `mazd//after` macro to replace occurrences of
`with-eval-after-load` across multiple modules, simplifying the
syntax by removing the need for quoted features. This makes the
codebase more consistent and easier to read. The macro is
implemented in `mazd-func.el` and is applied to several files,
including `mazd-lsp.el`, `mazd-buffer.el`, and others.